new Index Page

+++ app/src/main/java/maimeng/yodian/app/client/android/network/service/SkillService.java
@@ -35,7 +35,7 @@ public interface SkillService {
      * @param callback
      */
     @Post(Api.SKILL_CHOICE)
-    void choose(@Param("p") int p, @Param("scid") int scid, Callback<SkillResponse> callback);
+    void choose(@Param("p") int p, @Param("scid") long scid, Callback<SkillResponse> callback);
 
     /**
      * 获取技能模板
